refactor: move RawQueryResult to types layer, widen series data, extract sort comparator
- Move RawQueryResult from panelRegistry.ts to types/panel.ts so it
  lives alongside other panel-related types. Re-export from
  panelRegistry.ts for backward compatibility.
- Widen series[].data from Array<{ timestamp; value }> to unknown[]
  so future panel types (heatmap, histogram) can use different shapes.
- Extract duplicated localeCompare sort into a private byLabel constant.
